+++ NEWS.md
@@ -29,6 +29,7 @@
 - aggregate_election_data() function now includes a list of id_candidacies at province level when we aggregate 
 - Fixed summary_election_data()
 - Included a CERA flag in summary_election_data() to allow remove CERA ballots.
+- Included first unit tests for utils.R
 
 - pending: Fixed documentation of get_elections_data(), aggregate_election_data(), summary_election_data()
 - pending: add more examples in summary()